Put _get_threads_chunks in utils.py
