+++ pygbm/splitting.py
@@ -2,12 +2,14 @@
 import numpy as np
 from numba import njit, jitclass, prange, float32, uint8, uint32
 import numba
+
 from .histogram import _build_histogram
 from .histogram import _subtract_histograms
 from .histogram import _build_histogram_no_hessian
 from .histogram import _build_histogram_root
 from .histogram import _build_histogram_root_no_hessian
 from .histogram import HISTOGRAM_DTYPE
+from .utils import _get_threads_chunks
 
 
 @jitclass([
@@ -336,25 +338,14 @@ def find_node_split(context, sample_indices):
     #     ctx.ordered_gradients[i] = ctx.gradients[samples_indices[i]]
     # Ordering the gradients and hessians helps to improve cache hit.
     if sample_indices.shape[0] != ctx.gradients.shape[0]:
-        n_threads = numba.config.NUMBA_DEFAULT_NUM_THREADS
-        # Each threads writes data in ordered_xx from starts[thread_idx] to
-        # starts[thread_idx] + sizes[thread_idx]
-        sizes = np.full(n_threads, n_samples // n_threads, dtype=np.int32)
-        if n_samples % n_threads > 0:
-            # array[:0] will cause a bug in numba 0.41 so we need the if.
-            # Remove once issue numba 3554 is fixed.
-            sizes[:n_samples % n_threads] += 1
-        starts = np.zeros(n_threads, dtype=np.int32)
-        starts[1:] = np.cumsum(sizes[:-1])
+        starts, ends, n_threads = _get_threads_chunks(n_samples)
         if ctx.constant_hessian:
             for thread_idx in prange(n_threads):
-                for i in range(starts[thread_idx],
-                               starts[thread_idx] + sizes[thread_idx]):
+                for i in range(starts[thread_idx], ends[thread_idx]):
                     ordered_gradients[i] = ctx.gradients[sample_indices[i]]
         else:
             for thread_idx in prange(n_threads):
-                for i in range(starts[thread_idx],
-                               starts[thread_idx] + sizes[thread_idx]):
+                for i in range(starts[thread_idx], ends[thread_idx]):
                     ordered_gradients[i] = ctx.gradients[sample_indices[i]]
                     ordered_hessians[i] = ctx.hessians[sample_indices[i]]
